requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
idna==2.10
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
six==1.16.0
telegram-send==0.25
tornado==6.1
tzlocal==2.1
urllib3==1.26.5
